Analysis: Move SlowControl to own subdir
