+++ src/analysis/slowcontrol/processors/SlowControlProcessor.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace ant{
+namespace analysis {
+namespace slowcontrol {
+
+
+
+}
+}
+}
